refactor/move cursor
